+++ backend/src/main/java/codesquard/app/domain/item/ItemRepository.java
@@ -1,9 +1,9 @@
-package codesquard.app.domain.item.repository;
+package codesquard.app.domain.item;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import codesquard.app.domain.item.entity.Item;
+import codesquard.app.domain.item.Item;
 
 @Repository
 public interface ItemRepository extends JpaRepository<Item, Long> {
