+++ backend/src/main/java/codesquard/app/domain/member/Member.java
@@ -1,4 +1,4 @@
-package codesquard.app.domain.member.entity;
+package codesquard.app.domain.member;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
