+++ backend/src/main/java/codesquard/app/domain/member/MemberRepository.java
@@ -1,9 +1,9 @@
-package codesquard.app.domain.member.repository;
+package codesquard.app.domain.member;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import codesquard.app.domain.member.entity.Member;
+import codesquard.app.domain.member.Member;
 
 @Repository
 public interface MemberRepository extends JpaRepository<Member, Long> {
